feat: add route to like a sauce

+++ src/controllers/sauces.controller.js
@@ -17,7 +17,7 @@ exports.getSauce = async (req, res, next) => {
 
   try {
     const sauce = await Sauce.findById(id)
-    if (!sauce) throw createError(404, 'sauce not found')
+    if (!sauce) throw createError.NotFound()
 
     res.status(200).json(sauce)
   } catch (err) {
@@ -26,36 +26,43 @@ exports.getSauce = async (req, res, next) => {
 }
 
 exports.createSauce = async (req, res, next) => {
+  const { userId } = req
+
   try {
-    const sauce = new Sauce({
+    const data = {
       ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    })
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+      userId
+    }
+
+    const sauce = new Sauce(data)
     await sauce.save()
 
     res.status(200).send({ message: 'sauce created successfully' })
   } catch (err) {
-    await fs.unlink(req.file.path)
+    if (req.file) await fs.unlink(req.file.path)
     next(err)
   }
 }
 
 exports.updateSauce = async (req, res, next) => {
   const { id } = req.params
+  const { userId } = req
 
   try {
-    const sauce = await Sauce.findById(id)
-    if (!sauce) throw createError(404, 'sauce not found')
+    const data = req.file
+      ? {
+          ...JSON.parse(req.body.sauce),
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        }
+      : req.body
 
-    await sauce.updateOne(
-      req.file
-        ? {
-            ...JSON.parse(req.body.sauce),
-            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-          }
-        : req.body,
-      { runValidators: true }
-    )
+    const sauce = await Sauce.findById(id)
+    if (!sauce) throw createError.NotFound()
+
+    if (sauce.userId !== userId) throw createError.Forbidden()
+
+    await sauce.updateOne(data, { runValidators: true })
 
     if (req.file) {
       await fs.unlink(`public/images/${sauce.imageUrl.split('/').pop()}`)
@@ -70,16 +77,58 @@ exports.updateSauce = async (req, res, next) => {
 
 exports.deleteSauce = async (req, res, next) => {
   const { id } = req.params
+  const { userId } = req
 
   try {
     const sauce = await Sauce.findById(id)
-    if (!sauce) throw createError(404, 'sauce not found')
+    if (!sauce) throw createError.NotFound()
+
+    if (sauce.userId !== userId) throw createError.Forbidden()
 
     await sauce.deleteOne()
 
     await fs.unlink(`public/images/${sauce.imageUrl.split('/').pop()}`)
 
     res.status(200).json('sauce deleted successfully')
+  } catch (err) {
+    next(err)
+  }
+}
+
+exports.likeSauce = async (req, res, next) => {
+  const { id } = req.params
+  const { userId } = req
+  const { like } = req.body
+
+  try {
+    const sauce = await Sauce.findById(id)
+    if (!sauce) throw createError.NotFound()
+
+    let { usersLiked, usersDisliked } = sauce
+
+    switch (like) {
+      case 1:
+        usersLiked = usersLiked.includes(userId) ? usersLiked : [...usersLiked, userId]
+        usersDisliked = usersDisliked.filter((x) => x !== userId)
+        break
+      case -1:
+        usersDisliked = usersDisliked.includes(userId) ? usersDisliked : [...usersDisliked, userId]
+        usersLiked = usersLiked.filter((x) => x !== userId)
+        break
+      case 0:
+        usersLiked = usersLiked.filter((x) => x !== userId)
+        usersDisliked = usersDisliked.filter((x) => x !== userId)
+        break
+      default:
+        throw createError.BadRequest()
+    }
+
+    const likes = usersLiked.length
+    const dislikes = usersDisliked.length
+
+    await sauce.updateOne({ usersLiked, usersDisliked, likes, dislikes })
+
+    res.status(200).send({ message: 'sauce liked / disliked successfully' })
   } catch (err) {
     next(err)
   }
